Refactor: 게임방 현재 인덱스 증가 코드 수정, 회원 탈퇴 시  MemberRoles 변경 & MemberQuit에 추가

+++ src/main/java/com/garamgaebi/GaramgaebiServer/domain/ice_breaking/service/GameService.java
@@ -45,7 +45,7 @@ public class GameService {
         for (int i = 0; i < 8; i++) { // 방 8개씩 생성
             roomId = UUID.randomUUID().toString();
 
-            ProgramGameroom room = ProgramGameroom.builder().programIdx(programIdx).roomId(roomId).currentImgIdx(0L).build();
+            ProgramGameroom room = ProgramGameroom.builder().programIdx(programIdx).roomId(roomId).currentImgIdx(0).build();
 
             programGameroomRepository.save(room);
             rooms.add(room);
@@ -134,7 +134,7 @@ public class GameService {
         return images;
     }
 
-    public Long getCurrentImgIdx(String roomId) {
+    public int getCurrentImgIdx(String roomId) {
         ProgramGameroom room = programGameroomRepository.findByRoomId(roomId)
                 .orElseThrow(() -> new RestApiException(ErrorCode.NOT_EXIST_GAME_ROOM));
